requirements.txt
Flask
Flask-Cors
mysql-connector-python
bcrypt
PyJWT
python-dotenv
gunicorn
requests
python-dotenv